requirements.txt
Django
django-achilles
django-forms-bootstrap
six
python-ldap
djangorestframework
